Made it use a double

+++ Tourny.java
@@ -1,7 +1,7 @@
 class Tourny{
    public static void main(String args[]){
        long i =0L;
-       for (long peopleLeft = 7900000000L; peopleLeft > 1; peopleLeft /= 2){
+       for (double peopleLeft = 7900000000; peopleLeft > 1; peopleLeft /= 2){
            ++i;
            System.out.println(i+":"+peopleLeft);
        }
